requirements.txt
psycopg2-binary==2.9.9
streamlit==1.34.0
openai==1.22.0           # streaming & 4o-mini support
streamlit-audiorec==0.1.3
passlib==1.7.4
bcrypt==3.2.2
fpdf2>=2.6.1
arabic_reshaper>=3.0.0
python-bidi>=0.6.6
httpx==0.25.2
